feat: Refactor sleep aggregation to support multi-day sessions
This commit refactors the sleep data processing to correctly handle sleep sessions that span across multiple days. Previously, sleep was attributed only to its start date, leading to inaccurate daily and aggregated totals for overnight sleep.

The new implementation splits multi-day sleep sessions at midnight, attributing the correct duration to each calendar day. This ensures that daily, weekly, monthly, and yearly sleep aggregations are accurate.

Key changes:

-   **`SleepRepository.kt`:**
    -   Reworked `getSleepsBetween` to process sleep sessions that cross midnight. It now splits a single `SleepEntity` into multiple `Sleep` objects, one for each day it covers.
    -   The split segments are clipped to the boundaries of the queried date range.
    -   Added `SleepRepositoryTest.kt` with comprehensive unit tests to verify the new splitting logic, including cases for single-day sleep, overnight sleep, multi-day sleep, clipping at date range boundaries, and handling of Daylight Saving Time changes.

-   **`SleepOverviewViewModel.kt`:**
    -   Introduced `aggregateDailySleeps` to pre-process raw `Sleep` data, grouping and summing sleep durations per calendar day. This provides an accurate daily total before further aggregation into weeks, months, or years.
    -   Refactored `mapWeek`, `mapMonth`, and `mapYear` to consume the new pre-aggregated daily sleep data (`List<SleepOverview>`) instead of raw `List<Sleep>`.
    -   Updated the logic for calculating average sleep in `aggregateSleepsByPeriod` to correctly divide by the number of days in the period (e.g., 7 for a week, days in the month for a month), providing a more accurate average.
    -   Calculations now use `Double` for precision with minutes.

-   **`PlotSleeps.kt` & `SleepOverview.kt`:**
    -   Updated the `PlotSleeps` composable and related functions to use `Double` types for Y-axis models (`DoubleLinearAxisModel`) and plot data, aligning with the change to floating-point minute calculations in the ViewModel.
    -   Changed `SleepOverview.totalMinutes` from `Int` to `Double`.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/sleep/SleepOverview.kt
@@ -5,7 +5,7 @@ import org.darthacheron.fitbe.health.componenets.Overview
 
 data class SleepOverview(
     override val date: LocalDate,
-    val totalMinutes: Int,
+    val totalMinutes: Double,
 ) : Overview {
     val minutes = totalMinutes % 60
     val hours = totalMinutes / 60
